Cleaned up simulation
