+++ src/simulation/village_manager.rs
@@ -1,0 +1,9 @@
+use village::*;
+use village_mind::*;
+
+pub struct VillageManager {
+    pub village: VillageRef,
+    pub village_mind: VillageMind,
+}
+
+
