+++ src/village/resource.rs
